Project1_Fancy_Lighting
This fancy lighting is a lighting system controlled by two pushbuttons switch and a photo sensor. It has a pushbutton to controls its turning-on/off. It uses an RGB LED as the lighting. When light is dark, LED is turned on. When light is bright, LED is turned off. Meanwhile, there is a pushbutton controlling the color switch, including 6 colors defined as below.
#define NUM_COLORS 6 
float colors[NUM_COLORS][3] = {{1.0,0.0,0.0}, {1.0,1.0,0.0}, {0.0,1.0,0.0}, {0.0,1.0,1.0}, {0.0,0.0,1.0}, {1.0,0.0,1.0}}; 
The color switch is a smooth color changing. When changing from one color to another, it spend 2 seconds to change the color gradually and smoothly.